Refactor backend database integration: update test database import path, and clean up supabase client code
